Refactor protocol hook into its own file
